Create HTML recipe file for katsu dish
I created katsu.html that will be used to display a recipe page
for the Chicken Katsu dish. The current contents of the HTML file is
the boilerplate HTML set up and the first header. I also created a
link on the index.html file to the Chicken Katsu recipe page using an
anchor element. I implemented an unordered list element for the recipe
links to ensure they were displayed on separate lines.