Add ingredients and steps to katsu recipe
I used an unordered list element to list out the ingredients needed
for the chicken katsu recipe. I also used an ordered list element to
list out the steps required to make the chicken katsu dish. Adding
the ingredients and steps section to the katsu recipe page will make
it have the same structure as the pho recipe page for the sake of
consistency.