Add image and description to the gabli recipe page
I used an image element to add a picture of the korean dish
to the galbi recipe page. I also used a paragraph element to add a
description of the galbi dish on the recipe page. These edits will
make the structure of the galbi recipe page the same as the other recipe
pages.