Add ingredients and steps to galbi recipe page
I used an unordered list element to list out the required ingredients
for the galbi recipe. I used an ordered list element to list out
the steps for creating the galbi dish. Additionally, I added the
total time to complete the recipe and serving number details under the
description section of the recipe page. These changes make the galbi
recipe page identical in structure to the other recipe pages.