Add style to elements on galbi recipe page
I added the appropriate classes to the different text elements on the
galbi recipe page. I also added a ID to the png image to resize the
image in the styles.css file. I also made a few edits to the font sizes
of some classes and edited a few class names.