Edit recipe links and their respective classes
I turned the list elements for the recipe links into flex elements.
This allowed me to display the recipe links horizontally.
I also added a background and a ridge border to the recipe links to
add some visual appeal. I removed the info-text class in the index.html
file because it conflicted with my recipe-links class. Furthermore,
I increased the size of the recipe links font. The recipe links were
centered and evenly spaced.