Edit margins and padding of the informational text
The informational content on the recipe web page was edited so that
the borders of the content were further from the browser edge. This
should add more visual appeal because the informational text will
not stretch out too close to the edge of the browser. The second header
(small-header) top margins were increased so that it is easier to
distinguish between the different sections.