Edit image elements and add spacing under them
I removed the left margin for each food image because I used
flexbox to center them on the main axis (so left margin specification
was no longer needed). I also added a bottom margin under each food
image to make it easier to notice the different sections of the page
and to add visual appeal.